refactor(home): defer showcase client islands
